correção do erro do cálculo da distância

+++ features/nearby/src/main/java/com/example/nearby/domain/entity/NearbyStores.kt
@@ -4,5 +4,5 @@ internal data class NearbyStores(
     val id: Int,
     val name: String? = "",
     val iconUrl: String? = "",
-    val distance: Float
+    val distance: Int
 )
